pyproject.toml
[project]
name = "ia-reciclaje"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imutils~=0.5.4",
    "opencv-python~=4.10.0.84",
    "numpy~=2.1.2",
    "pillow~=10.4.0",
    "ultralytics~=8.3.13",
    "requests>=2.32.3",
]
